+++ src/main/java/com/github/upcraftlp/respawnlocationpicker/api/package-info.java
@@ -3,7 +3,7 @@
  */
 @MethodsReturnNonnullByDefault
 @ParametersAreNonnullByDefault
-@API(apiVersion = "1.0.0", owner = "respawn-location-picker", provides = "respawn-location-picker-API")
+@API(apiVersion = "1.1.0", owner = "respawn-location-picker", provides = "respawn-location-picker-API")
 package com.github.upcraftlp.respawnlocationpicker.api;
 
 import mcp.MethodsReturnNonnullByDefault;