+++ src/main/java/com/github/upcraftlp/respawnlocationpicker/client/gui/GuiRespawnLocations.java
@@ -1,12 +1,13 @@
 package com.github.upcraftlp.respawnlocationpicker.client.gui;
 
+import com.github.upcraftlp.respawnlocationpicker.api.util.TargetPoint4d;
 import com.github.upcraftlp.respawnlocationpicker.net.NetworkHandler;
 import com.github.upcraftlp.respawnlocationpicker.net.packet.PacketSetRespawnLocation;
-import com.github.upcraftlp.respawnlocationpicker.util.TargetPoint4d;
 import net.minecraft.client.gui.GuiButton;
 import net.minecraft.client.gui.GuiScreen;
 import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.client.resources.I18n;
+import net.minecraft.util.NonNullList;
 import net.minecraft.util.text.TextFormatting;
 import net.minecraftforge.fml.client.config.GuiButtonExt;
 
@@ -17,34 +18,37 @@ import java.io.IOException;
  */
 public class GuiRespawnLocations extends GuiScreen {
 
-    private final TargetPoint4d[] targets;
-    private String[] labels;
+    private final NonNullList<TargetPoint4d> targets;
+    private final boolean hasHoverText;
+    private NonNullList<String> labels = null;
     private static final int
             BUTTON_HEIGHT = 20,
             MARGIN = 8;
     private int BUTTON_WIDTH = 100;
 
 
-    public GuiRespawnLocations(TargetPoint4d[] targets) {
+    public GuiRespawnLocations(NonNullList<TargetPoint4d> targets, boolean showHoverText) {
         this.targets = targets;
+        this.hasHoverText = showHoverText;
+        if(showHoverText) labels = NonNullList.create();
     }
 
     @Override
     public void initGui() {
         int x = width / 2;
         int y = width / 10 + 10;
-        labels = new String[this.targets.length];
 
         //make sure to adapt the size of the buttons to the label widths
         for(TargetPoint4d target : this.targets) {
             int width = mc.fontRenderer.getStringWidth(target.getName()) + (MARGIN * 2);
             if(width > BUTTON_WIDTH) BUTTON_WIDTH = width;
-
         }
-        for(int i = 0; i < this.targets.length; i++) {
-            labels[i] = getStringForTarget(targets[i], true);
+        for(int i = 0; i < this.targets.size(); i++) {
+            if(hasHoverText) {
+                labels.add(getStringForTarget(targets.get(i), true));
+            }
             int xPos = x - (BUTTON_WIDTH + 5) + ((i % 2) * (BUTTON_WIDTH + 10));
-            this.buttonList.add(new GuiButtonExt(i, xPos, y + (i / 2) * (BUTTON_HEIGHT + 5), BUTTON_WIDTH, BUTTON_HEIGHT, targets[i].getName()));
+            this.buttonList.add(new GuiButtonExt(i, xPos, y + (i / 2) * (BUTTON_HEIGHT + 5), BUTTON_WIDTH, BUTTON_HEIGHT, targets.get(i).getName()));
         }
     }
 
@@ -56,12 +60,13 @@ public class GuiRespawnLocations extends GuiScreen {
         GlStateManager.scale(2.0F, 2.0F, 2.0F);
         this.drawCenteredString(this.fontRenderer, I18n.format("gui.respawnLocation.title"), this.width / 2 / 2, 15, 16777215);
         GlStateManager.popMatrix();
-        for (GuiButton button : this.buttonList) {
-            if(button.isMouseOver()) {
-                drawHoveringText(labels[button.id], mouseX, mouseY);
+        if(hasHoverText) {
+            for (GuiButton button : this.buttonList) {
+                if(button.isMouseOver()) {
+                    drawHoveringText(labels.get(button.id), mouseX, mouseY);
+                }
             }
         }
-
     }
 
     @Override
@@ -71,24 +76,31 @@ public class GuiRespawnLocations extends GuiScreen {
     }
 
     private static String getStringForTarget(TargetPoint4d target, boolean includeFormatting) {
-        return
-                //DIMENSION
-                (includeFormatting ? TextFormatting.DARK_AQUA : "") +
-                "DIM " +
-                target.getDimension() +
-                ":" +
-                (includeFormatting ? TextFormatting.RESET : "") +
+        StringBuilder builder = new StringBuilder();
+        if(target.hasPosition()) {
+            builder
+                    //DIMENSION
+                    .append(includeFormatting ? TextFormatting.DARK_AQUA : "")
+                    .append("DIM ").append(target.getDimension()).append(":")
+                    .append(includeFormatting ? TextFormatting.RESET : "")
+                    .append(" ")
 
-                //SPAWNPOINT LOCATION
-                " " +
-                (includeFormatting ? TextFormatting.DARK_PURPLE : "") +
-                "[" +
-                target.getX() +
-                ", " +
-                target.getY() +
-                ", " +
-                target.getZ() +
-                "]";
+                    //POSITION
+                    .append(includeFormatting ? TextFormatting.DARK_PURPLE : "")
+                    .append("[")
+                    .append(target.getX())
+                    .append(", ")
+                    .append(target.getY())
+                    .append(", ")
+                    .append(target.getZ()).append("]");
+        }
+        if(target.hasBiome()) {
+            builder
+                    .append("(")
+                    .append(target.getBiome())
+                    .append(")");
+        }
+        return builder.toString();
     }
 
 }