+++ changelog.md
@@ -1,4 +1,11 @@
 ------------------------------------------------------
+Version 1.1.0
+------------------------------------------------------
+- fixed possible client crash
+- added hook for the vanilla /spawnpoint command
+- improved compatibility with other mods
+
+------------------------------------------------------
 Version 1.0.0
 ------------------------------------------------------
 - initial release